Created a super class for three classes, added a module for teacher and apprentice_teacher classes
