+++ apprentice_teacher.rb
@@ -1,6 +1,15 @@
-class ApprenticeTeacher
+require_relative 'at_school'
+require_relative 'teachable'
+
+class ApprenticeTeacher < At_School
   attr_reader :age, :salary, :phase, :target_raise
   attr_accessor :name
+
+  PHASE = 2
+  TARGET_RAISE = 800
+  RATING = 80
+  RESPONSE = "Listen, class, this is how everything works. *drops crazy knowledge bomb* ... You're welcome."
+
 
   def initialize(options={})
     @age = options.fetch(:age, 0)
@@ -9,42 +18,45 @@ class ApprenticeTeacher
     @phase = 3
   end
 
-  def offer_high_five
-    "High five!"
-  end
+  include Teachable
+
+  # def offer_high_five
+  #   "High five!"
+  # end
 
   def set_phase(num)
     @phase = num
     "Cool, I've always wanted to teach phase #{num}!"
   end
 
-  def teach_stuff
-    response = ""
-    response += "Listen, class, this is how everything works. "
-    response += "*drops crazy knowledge bomb* "
-    response += "... You're welcome."
-    response
-  end
+  # def teach_stuff
+  #   # response = ""
+  #   # response += "Listen, class, this is how everything works. "
+  #   # response += "*drops crazy knowledge bomb* "
+  #   # response += "... You're welcome."
+  #   # response
+  #   RESPONSE
+  # end
 
-  def salary=(new_salary)
-    puts "This better be good!"
-    @salary = new_salary
-  end
+  # def salary=(new_salary)
+  #   puts "This better be good!"
+  #   @salary = new_salary
+  # end
 
-  def receive_raise(raise)
-    @salary += raise
-  end
+  # def receive_raise(raise)
+  #   @salary += raise
+  # end
 
-  def set_performance_rating(rating)
-    response = ""
-    if rating > 80
-      response = "Yay, I'm a great employee!"
-      receive_raise(@target_raise)
-    else
-      response = "Oh, well -- thanks to this actionable, specific, and kind feedback, I'll do better next time."
-    end
-    response
-  end
+  # def set_performance_rating(rating)
+  #   response = ""
+  #   if rating > 80
+  #     response = "Yay, I'm a great employee!"
+  #     receive_raise(@target_raise)
+  #   else
+  #     response = "Oh, well -- thanks to this actionable, specific, and kind feedback, I'll do better next time."
+  #   end
+  #   response
+  # end
 
   def attend_training_session
     puts "Whoa. I know ruby-fu"
