+++ teachable.rb
@@ -1,0 +1,39 @@
+module Teachable
+  # def teach_stuff
+  #   response = ""
+  #   response += "Listen, class, this is how everything works. "
+  #   response += "*drops crazy knowledge bomb* "
+  #   response += "... You're welcome."
+  #   response
+  # end
+
+  def salary=(new_salary)
+    puts "This better be good!"
+    @salary = new_salary
+  end
+
+  def receive_raise(raise)
+    @salary += raise
+  end
+
+  def set_performance_rating(rating)
+    response = ""
+    if rating > self.class::RATING
+      receive_raise(@target_raise)
+      response = "Yay, I'm a great employee!"
+    else
+      response += "Oh, well -- thanks to this actionable, specific, and kind "
+      response += "feedback, I'll do better next time."
+    end
+    response
+  end
+
+  def teach_stuff
+    # response = ""
+    # response += "Listen, class, this is how everything works, fo SHO! "
+    # response += "*drops flat-out insane knowledge bomb* "
+    # response += "... You're welcome. *saunters away*"
+    # response
+    self.class::RESPONSE
+  end
+end
